Revert "Implement attachLabelClickHandlers utility for dynamic label click events"
This reverts commit 55d84d70509901702edd3aac5f7c3a5e0320a386.
