+++ components/interface/utils/utils.js
@@ -433,45 +433,10 @@ var hasVisualType = function (variableId) {
   }
 };
 
-var attachLabelClickHandlers = function () {
-  // Select all label tags
-  const labelElements = document.querySelectorAll('.label.types > .label[class*="label-"]');
-  
-  labelElements.forEach(label => {
-    // Avoid attaching multiple handlers
-    if (!label.dataset.handlerAttached) {
-      label.dataset.handlerAttached = 'true';
-      
-      // Extract label type from class name
-      const classNames = Array.from(label.classList);
-      const labelClass = classNames.find(cls => cls.startsWith('label-'));
-      
-      if (labelClass) {
-        const labelType = labelClass.replace('label-', '');
-        
-        // Attach click handler
-        label.addEventListener ('click', event => {
-          event.stopPropagation();
-          const termID = labelTypeToID[labelType];
-          
-          if (termID) {
-            if (window.Instances && window.Instances.getInstance(termID)) {
-              window.setTermInfo(window.Instances.getInstance(termID)[termID + "_meta"], termID);
-            } else {
-              window.addVfbId(termID);
-            }
-          }
-        });
-      }
-    }
-  });
-}
-
 module.exports = {
   setSepCol,
   getStackViewerDefaultX,
   getStackViewerDefaultY,
   hasVisualType,
-  labelTypeToID,
-  attachLabelClickHandlers
+  labelTypeToID
 };
